Removed non-POSIX bashism from test operation
The [ command (a.k.a. test) must be used carefully in portable scripts.
A very common pitfall is to use the == operator, which is a bashism;
use = instead.

Some users reported problems when running this script under Ubuntu
and Debian.  It seems that these distros use dash as their default
shell.  dash would silently fail due to non-POSIX commands in the
script.

To find the non-POSIX commands, I used the 'checkbashism' script
from http://sourceforge.net/projects/checkbaskisms.

This script is recommended in: https://wiki.ubuntu.com/DashAsBinSh

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>