Fixed quoting of files paths in targeted diff
When trying to the limit the scope of diff to a set of files
(e.g. git diffall -- foo bar), the script was incorrectly placing
quotes around the git invocations.  The quotes were interfering with
the filename filter on diff, so they needed to be removed.

This fixes GitHub Issue #17

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>