Cleaned up trap function so it actually runs on abnormal exit
In earlier versions, trap was not set until the last line of the script.
This prevented it from executing on any abnormal exits.  It must be set
up *before* it is needed.  This version also moves the '--copy-back'
feature to the end of normal execution rather than making it part of the
exit trap.

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>