Verify that user supplied a command for --extcmd
If the user executed "git diffall --extcmd", it would result in a shift
error with no useful message to the user.

The script now verifies that a command was specified after --extcmd if
it is the last option on the command-line.  If it fails, an error
message and usage instructions are displayed.

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>