fix handling of submodules that live in subdirectories
If the diff includes a submodule that lives in a subdirectory of
the project, the script fails with "No such file or directory".

For example, if a submodule is located at '<repo root>/lib/sub',
the script fails when it tries to write the submodule SHA1 to
'/tmp/.../lib/sub' because the 'lib' directory does not exist.

This commit teaches diffall to create the subdirectory path in the
tmp location prior to writing the submodule file.

This closes #21 on GitHub

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>