append "dirty" to submodule SHA1 when needed
If a submodule is dirty (i.e. has uncommitted changes) then

  'git diff' reports "<SHA1 of parent>-dirty"

while

  'git diff --raw' reports "<SHA1 of parent>"

diffall uses the '--raw' diff format, so the "dirty" flag is lost.

To make the output more closely match that of 'git diff', teach
diffall to append "-dirty" to the submodule SHA1 when needed.

This closes #22 on GitHub.

Signed-off-by: Tim Henigan <tim.henigan@gmail.com>